+++ crate_anon/tools/print_crateweb_demo_config.py
@@ -189,6 +189,7 @@ DATABASES = {
     # Optional: 'cpft_rio_datamart'
     # Optional: 'cpft_rio_raw'
     # Optional: 'cpft_rio_rcep'
+    # Optional: 'cpft_systmone'
     # ... see ClinicalDatabaseType in crate_anon/crateweb/config/constants.py
 }